pyproject.toml
[project]
name = "my-adk-expert-agent"
version = "0.1.0"
description = "ADK Expert Bot"
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.0.0",
    "browser-use", # Keep if other parts still use it
    "langchain-google-genai",
    "google-cloud-secret-manager",
    "python-dotenv",
    "google-cloud-storage", # Added dependency
    "pydantic>=2.0" # Ensure pydantic is explicit if needed for Literal, etc.
]

[project.scripts]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.11.13",
]
